# services/registry/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Users, roles, authentication. Phase A: admin-created users, activation via\ntoken, login, password change. Sending real email (invites/forgot-password)\nis a later phase — see docs/adr/0013.",
        "title": "Sanitas — Registry API",
        "contact": {},
        "version": "0.1.0"
    },
    "paths": {
        "/.well-known/jwks.json": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Public key to verify JWTs (JWKS format)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/healthz": {
            "get": {
                "tags": [
                    "system"
                ],
                "summary": "Liveness check",
                "responses": {
                    "200": {
                        "description": "Service healthy"
                    },
                    "503": {
                        "description": "Database unreachable"
                    }
                }
            }
        },
        "/v1/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Email or username + password",
                        "name": "credenziali",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.authTokens"
                        }
                    },
                    "400": {
                        "description": "Invalid payload"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/v1/logout": {
            "post": {
                "description": "Invalidates the given refresh token. An access token already issued stays\nvalid until its own expiry (24h) — no revocation mechanism for those yet,\nsee docs/adr/0013.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "refresh",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.refreshRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Logged out"
                    },
                    "400": {
                        "description": "Invalid payload"
                    },
                    "401": {
                        "description": "Invalid, expired, or already used token"
                    }
                }
            }
        },
        "/v1/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get the authenticated user's profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_francescocerri_sanitas_services_registry_internal_user.User"
                        }
                    },
                    "401": {
                        "description": "Authentication required"
                    }
                }
            }
        },
        "/v1/password/change": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Change the authenticated user's password",
                "parameters": [
                    {
                        "description": "Old and new password",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.changePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Password changed"
                    },
                    "400": {
                        "description": "Invalid payload"
                    },
                    "401": {
                        "description": "Old password incorrect"
                    }
                }
            }
        },
        "/v1/refresh": {
            "post": {
                "description": "Consumes the given refresh token (single use — see docs/adr/0016) and\nreturns a new access token plus a new refresh token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh the access token",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "refresh",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.refreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.authTokens"
                        }
                    },
                    "400": {
                        "description": "Invalid payload"
                    },
                    "401": {
                        "description": "Invalid, expired, or already used token"
                    }
                }
            }
        },
        "/v1/users": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a new user (requires the users:manage permission)",
                "parameters": [
                    {
                        "description": "Email, username, roles (slug)",
                        "name": "utente",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.createUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.createUserResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payload or unknown role"
                    },
                    "403": {
                        "description": "Missing required permission: users:manage"
                    },
                    "409": {
                        "description": "Email or username already in use"
                    }
                }
            }
        },
        "/v1/users/activate": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Activate an account via the invite token, setting the password",
                "parameters": [
                    {
                        "description": "Invite token and new password",
                        "name": "attivazione",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.activateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Account activated"
                    },
                    "400": {
                        "description": "Invalid payload"
                    },
                    "401": {
                        "description": "Invalid, expired, or already used token"
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_francescocerri_sanitas_services_registry_internal_user.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.activateUserRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.authTokens": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.changePasswordRequest": {
            "type": "object",
            "properties": {
                "new_password": {
                    "type": "string"
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.createUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.createUserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invite_url": {
                    "description": "InviteURL is handed back directly in this response instead of being\nemailed: sending it for real (Gmail SMTP) is a separate follow-up\nactivity, this phase only lays the token groundwork. The admin\nforwards it to the volunteer by whatever channel they already use.",
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.loginRequest": {
            "type": "object",
            "properties": {
                "identifier": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.refreshRequest": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the access token from POST /v1/login (e.g. \"Bearer eyJhbGci...\").",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}